Moved path operations for posts and users into seperate files and all corresponding changes. Tested each operation success and failures in Postman.
