+++ Learning/FastAPI/app/routers/post.py
@@ -1,0 +1,64 @@
+# Path operations converning posts
+from fastapi import Body, FastAPI, Response, status, HTTPException, APIRouter
+from app import schema as sch;
+from app.database import get_db;
+
+router = APIRouter()
+
+
+# Get all of the posts from the database
+@router.get("/posts")
+def get_posts():
+    conn, cursor = get_db()
+
+    cursor.execute("""SELECT * FROM posts""")
+    posts = cursor.fetchall()
+    return {"data": posts}
+
+# Create a brand new post
+@router.post("/posts", status_code=status.HTTP_201_CREATED)
+def create_posts(post: sch.Post):
+    conn, cursor = get_db()
+
+    cursor.execute("""INSERT INTO posts (title, content, published) VALUES (%s, %s, %s) RETURNING *""", (post.title, post.content, post.published))
+    new_post = cursor.fetchone()
+    conn.commit()   # changes made to the database must be committed deliberately
+    return {"data": new_post}
+
+# Get a single post based on the passed id
+@router.get("/posts/{id}")
+def get_post(id: int):
+    conn, cursor = get_db()
+
+    cursor.execute("""SELECT * FROM posts WHERE id = %s""", (str(id),))
+    post = cursor.fetchone()
+    if not post:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"post with id: {id} was not found")
+    return {"data": post}
+
+# Delete a post based on the passed id
+@router.delete("/posts/{id}")
+def delete_post(id:int):
+    conn, cursor = get_db()
+
+    cursor.execute("""DELETE FROM posts WHERE id = %s RETURNING *""", (str(id),))
+    deleted = cursor.fetchone()
+    if not deleted:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"post with id: {id} was not found")
+    conn.commit()   # deletion changes the database so it needs to be committed
+    return Response(status_code=status.HTTP_204_NO_CONTENT)
+
+# Update a post based on id
+@router.put("/posts/{id}")
+def update_post(id: int, post: sch.Post):
+    conn, cursor = get_db()
+    
+    cursor.execute("""UPDATE posts SET title = %s, content = %s, published = %s WHERE id = %s RETURNING *""", (post.title, post.content, post.published, str(id),))
+    updated = cursor.fetchone()
+    if not updated:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"post with id: {id} was not found")
+    conn.commit()
+    return {"data": updated}
